Reintroduce units
As of owlready version 2, [there are no defaults for data properties](https://owlready2.readthedocs.io/en/v0.37/porting1.html#functional-properties), and before we were _instantiating_ generics to use universally with some default. So as far as I can tell, in this class-based approach I'm left with no way to set a default value for the units (nor symbols) property(ies). So screw it. We can always to do point-to-point conversions, and in the meantime let's just add units as a column in the dataframe when scraping the pyiron database.